Move SvgSelectorDialog into symbology folder
